+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-None.
+- Implement `Stream` for `WebSocket`.
+- Implement `Sink` for `WebSocket`.
 
 ## Breaking changes